+++ BluebirdPS/private/Write-TwitterResponse.ps1
@@ -7,23 +7,11 @@ function Write-TwitterResponse {
 
     try {
 
-        if ($ResponseData.RateLimitRemaining -eq 0) {
-            $RateLimitReached = 'Rate limit of {0} has been reached. Please wait until {1} before making another attempt for this resource.' -f $ResponseData.RateLimit,$ResponseData.RateLimitReset
-            $RateLimitReached | Write-Error -ErrorAction Stop
-        }
-
-        if (($ResponseData.RateLimitRemaining -le $BluebirdPSConfiguration.RateLimitThreshold -and $null -ne $ResponseData.RateLimitRemaining)) {
-            $RateLimitMessage = 'The rate limit for this resource is {0}. There are {1} remaining calls to this resource until {2}. ' -f $ResponseData.RateLimit, $ResponseData.RateLimitRemaining, $ResponseData.RateLimitReset
-            switch ($BluebirdPSConfiguration.RateLimitAction) {
-                0 { $RateLimitMessage | Write-Verbose -Verbose; break}
-                1 { $RateLimitMessage | Write-Warning -Warning; break}
-                2 { $RateLimitMessage | Write-Error ; break}
-            }
-        }
-
         $BluebirdPSHistoryList.Add($ResponseData)
         Write-Information -MessageData $ResponseData
-        $script:LastResponseData = $ResponseData
+
+        [SuppressMessage('PSUseDeclaredVarsMoreThanAssigments', '')]
+        $global:BluebirdPSLastResponse = $ResponseData
 
         switch ($BluebirdPSConfiguration.OutputType) {
             'PSCustomObject' {